'完成了商品分类的增删改查'

+++ database/migrations/2017_04_20_155938_create_goods_table.php
@@ -16,9 +16,9 @@ class CreateGoodsTable extends Migration
         Schema::create('goods', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name', 30)->comment('商品名称');
-            $table->float('price', 8, 2)->index()->comment('商品价格');
+            $table->decimal('price', 8, 2)->index()->comment('商品价格');
             $table->text('description')->nullable()->comment('商品描述');
-            $table->float('promote_price')->nullable()->comment('促销价格');
+            $table->decimal('promote_price')->nullable()->comment('促销价格');
             $table->timestamp('promote_start_at')->nullable()->index()->comment('促销开始时间');
             $table->timestamp('promote_stop_at')->nullable()->index()->comment('促销结束时间');
             $table->unsignedTinyInteger('sort')->default(100)->index()->comment('权重');
